In contlog_encode_bounds, relax the conditions for speculatively flipping the lo bits, so that it can happen whenever the unknown value is known to be greater than 1/4.  This means that a "1..." may have to be corrected to a "01..." more frequently, but also that an iteration might be saved.
In contlog_encode_exact, handle the case when the result is > 1 and
there are some arg bits set, as might happen as a result of the change
described in the previous paragraph.  If the value is > 1, then undo
the bit flip from contlog_encode_bounds, do a shift, and then do a bit
flip.

This was found to be necessary in the 64-bit version of calculating
1/47 + 19/211.